requirements.txt
numpy==1.24.3
tensorboard==2.12.0
torch==2.4.1
torchsummary==1.5.1
torchvision==0.19.1
tqdm==4.66.4
omegaconf==2.3.0
hydra-core==1.3.2
